@@ -15,45 +15,47 @@
     "autogypi": "^0.2.2",
     "axios": "^0.17.1",
     "babel-plugin-styled-components": "^1.5.1",
-    "better-queue": "^3.8.6",
+    "better-queue": "^3.8.7",
     "codemirror": "^5.37.0",
     "concurrently": "^3.5.1",
-    "d3": "^5.1.0",
-    "electron": "^1.8.2",
-    "electron-dl": "^1.10.0",
+    "electron": "^1.8.6",
+    "electron-dl": "^1.12.0",
     "express": "^4.16.2",
     "gl-react": "^3.15.0",
     "gl-react-dom": "^3.15.0",
     "history": "4.7.2",
     "jsonfile": "^4.0.0",
     "keymaster": "^1.6.2",
+    "korg-nano-kontrol": "^0.2.9",
     "marked": "^0.3.6",
-    "mathjs": "^4.0.0",
+    "mathjs": "^4.2.2",
     "mobx": "^3.4.1",
     "mobx-react": "^4.3.5",
     "nanotimer": "^0.3.15",
     "nbind": "^0.3.14",
     "node-gyp": "^3.6.2",
-    "osc": "^2.2.0",
+    "osc": "^2.2.1",
+    "prismjs": "^1.14.0",
     "promise.prototype.finally": "^3.0.0",
     "prop-types": "^15.5.10",
     "query-string": "5.0.0",
-    "react": "^16.2.0",
+    "react": "^16.3.2",
     "react-app-rewire-mobx": "^1.0.8",
     "react-app-rewired": "^1.5.2",
     "react-codemirror2": "^3.0.7",
     "react-contextmenu": "^2.9.2",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.3.2",
     "react-draggable": "^3.0.5",
-    "react-grid-layout": "^0.16.2",
+    "react-grid-layout": "^0.16.6",
     "react-router-dom": "4.2.2",
     "react-scripts": "1.1.1",
     "react-selectable": "^2.0.0",
-    "reactjs-popup": "^1.0.4",
+    "reactjs-popup": "^1.1.1",
     "styled-components": "3.1.6",
-    "superagent": "^3.6.0",
+    "superagent": "^3.8.3",
     "superagent-promise": "^1.1.0",
     "supercolliderjs": "^0.14.2",
+    "tree-model": "^1.0.7",
     "wait-on": "^2.1.0"
   },
   "scripts": {
