@@ -4,24 +4,47 @@ import _ from 'lodash';
 
 import { save, timer } from '../keyFunctions'
 
+// import Prism from 'prismjs'
+
+import channelStore from "../stores/channelStore";
+import '../styles/Cell.css'
+
 @inject('cellStore')
 @observer
 export default class Cell extends React.Component {
     
     handleKeys = (event, channel_index, cell_index) => {
+        // 8 -> backspace
+        // 46 -> del
+        // 27 esc
 
+        // start timer
         if((event.metaKey || event.ctrlKey) && event.keyCode === 13) {
-            event.preventDefault();
             timer();
+            event.preventDefault();
         }
 
         //compile pattern with the parsed parameters + channel props cmd + enter
         else if(event.altKey && event.keyCode === 13) {
             this.props.cellStore.compileCell();
+            event.preventDefault();            
         }
 
+        // deselect with esc    
+        else if (event.keyCode === 27) {
+            this.props.cellStore.updateSelectState(false);
+            let elem = document.getElementById('cell_'+this.props.cellStore.current_cell[0]+"_"+
+                                                       this.props.cellStore.current_cell[1]);
+            if (elem)
+                elem.focus();
+            this.props.cellStore.selectCell(channel_index, cell_index);
+            this.nameInput.readOnly = false;
+
+            event.preventDefault();            
+        }
+        
         //select with enter
-        else if(event.keyCode === 13) {
+        else if (event.keyCode === 13) {
             if(!this.props.cellStore.isSelected) {
                 this.props.cellStore.updateSelectState(true);
                 this.props.cellStore.selectCell(channel_index, cell_index);
@@ -29,8 +52,11 @@ export default class Cell extends React.Component {
             }
             else {
                 this.props.cellStore.updateSelectState(false);
-                document.getElementById('cell'+this.props.cellStore.current_cell[0]+
-                                                this.props.cellStore.current_cell[1]).focus();
+                let elem = document.getElementById('cell_'+this.props.cellStore.current_cell[0]+"_"+
+                    this.props.cellStore.current_cell[1]);
+                
+                if (elem)
+                    elem.focus();
                 this.props.cellStore.selectCell(channel_index, cell_index);
                 this.nameInput.readOnly = false;
             }
@@ -50,15 +76,15 @@ export default class Cell extends React.Component {
                 this.props.cellStore.pasteCells();
             }
         }
-        //cut Cells
-        else if( (event.metaKey || event.ctrlKey)&& event.keyCode === 88 ) {
+        // cut Cells
+        else if( (event.metaKey || event.ctrlKey)&& event.key === 'x' ) {
             if(this.props.cellStore.isSelected) {
                 this.props.cellStore.cutCells();
                 event.preventDefault();
             }
         }
-        //delete selected cells
-        else if(event.keyCode === 8 && (event.metaKey || event.ctrlKey) ) {
+        //delete selected cells (backspace or del)
+        else if((this.props.cellStore.isSelected && (event.keyCode === 8 || event.keyCode === 46) )) {
             this.props.cellStore.deleteSelectedCells();
             event.preventDefault();
         }
@@ -106,37 +132,55 @@ export default class Cell extends React.Component {
     }
 
     render() {
-        console.log('RENDER CELL');
+        // console.log('RENDER CELL');
 
-        const item = this.props.item;
         const value = this.props.value;
         const cell_index = this.props.index;
         const channel_index = this.props.channel_index;
 
         var className = "GridItem";
-        if(this.props.cellStore.isCellActive(item.name, cell_index)){
-            className += ' active';
-        }
         className += (cell_index % 2 === 0) ? ' even' : ' odd'; 
-        if(this.props.cellStore.isCellSelected(channel_index, cell_index)) {
-            className += ' selected';
-        }
-        if(this.props.cellStore.isCellHighlighted(channel_index, cell_index)) {
-            className += ' highlighted';
-        }
+        
+        // eslint-disable-next-line no-unused-expressions
+        this.props.cellStore.updateCellSelectedClasses;
+
+        if (channelStore.getActiveChannels[channel_index].time %
+            channelStore.getActiveChannels[channel_index].steps === cell_index)
+            className += ' active'; 
+
         return (<div>
-            <textarea id={'cell'+channel_index+cell_index}
+            <textarea id={'cell_' + channel_index + "_" + cell_index}
                 ref={(input) => { this.nameInput = input; }}
                 className={className +" draggableCancel"} type="text"
                 placeholder={cell_index % 2 === 1 ? _.toString(cell_index+1) : ''}
                 value={value}
                 onChange={() => 
-                    (this.props.cellStore.updateCell(item.name, cell_index, this.nameInput.value))}
+                    (this.props.cellStore.updateCellValue(channel_index, cell_index, this.nameInput.value))}
                 onKeyDown={(event) => (this.handleKeys(event, channel_index, cell_index))}
                 onClick={() => {
                     this.props.cellStore.updateSelectState(false);
                     this.nameInput.focus();
                 }}/>
             </div>);
+
+        // return (<div className={"editor-holder"}>
+        //     <div className={"scroller"}>    
+        //         <textarea
+        //             className={"GridItem-core draggableCancel"}    
+        //             id={'cell_' + channel_index + "_" + cell_index}
+        //             ref={(input) => { this.nameInput = input; }}
+        //             placeholder={cell_index % 2 === 1 ? _.toString(cell_index + 1) : ''}
+        //             value={value}
+        //             onInput={() =>
+        //                 (this.props.cellStore.updateCellValue(item.name, cell_index, this.nameInput.value))}
+        //             onKeyDown={(event) => (this.handleKeys(event, channel_index, cell_index))}
+        //             onClick={() => {
+        //                 this.props.cellStore.updateSelectState(false);
+        //                 this.nameInput.focus();
+        //             }}/>
+        //         <pre><code className={className + " draggableCancel"}
+        //             dangerouslySetInnerHTML={{ __html: Prism.highlight(value, Prism.languages.javascript, 'javascript') }}></code></pre>
+        //     </div>
+        // </div>);
     }
 }
