@@ -1,20 +1,15 @@
 import _ from 'lodash';
 import React from 'react';
 import { inject, observer } from 'mobx-react';
+import { executionCssByEvent } from '../keyFunctions';
 
 @inject('sceneStore')
 @observer
 export default class Scene extends React.Component {
-    executionCss = (event, duration = 500) => {
-        event.persist();
-        event.target.className += ' Executed';
-        _.delay( () => (event.target.className = _.replace(event.target.className, ' Executed', '') ),
-                duration);
-    }
-
+    
     handleControlEnter = (event) => {
         if(event.ctrlKey && event.keyCode === 13){
-            this.executionCss(event);
+            executionCssByEvent(event);
             this.props.sceneStore.addScene(document.getElementById('new_scene_input').value)
         }
     }
@@ -34,7 +29,7 @@ export default class Scene extends React.Component {
     render() {
         console.log("RENDER SCENE.JS");
         
-        return (<div>
+        return (<div className={'Scenes PanelAdjuster'}>
             <input className={'Input draggableCancel'} id={"new_scene_input"}
                 placeholder={'New Scene Name'}
                 onKeyUp={this.handleControlEnter.bind(this)}/>
@@ -46,9 +41,14 @@ export default class Scene extends React.Component {
                 <button className={'Button draggableCancel'} 
                         onClick={() => (this.props.sceneStore.clearActiveGrid())}>Clear</button>
             </div>
+            <button className={'Button ' + (this.props.sceneStore.scene_mode ? 'true' : '') + ' draggableCancel'} 
+                style={{ width: '100%', height: '10px' }}    
+                title={'Song Mode: ' + (this.props.sceneStore.scene_mode ? 'On' : 'Off')}
+                onClick={() => (this.props.sceneStore.toggleScenemode())}> </button>
+            
             <div className={'AllScenes'}>
                 <div>
-                    {_.map(this.props.sceneStore.scene_list, this.renderScene.bind(this))}
+                    {_.map(this.props.sceneStore.scenesReversedOrder, this.renderScene.bind(this))}
                 </div>
             </div>
         </div>
