@@ -10,21 +10,20 @@ import './index.css';
 import Home from './components/Home';
 import MenuBar from './components/MenuBar';
 
-import layoutStore from './stores/layoutStore';
-import patternStore from './stores/patternStore';
-import consoleStore from './stores/consoleStore';
-import sceneStore from './stores/sceneStore';
-import menubarStore from './stores/menubarStore';
-import pulseStore from './stores/pulseStore';
-import pathStore from './stores/pathStore';
-import channelStore from './stores/channelStore';
-import rollStore from './stores/rollStore';
-import historyStore from './stores/historyStore';
-import globalStore from './stores/globalStore';
 import cellStore from './stores/cellStore';
-import nanoStore from './stores/nanoStore';
+import channelStore from './stores/channelStore';
+import consoleStore from './stores/consoleStore';
 import debugStore from './stores/debugStore';
-
+import globalStore from './stores/globalStore';
+import historyStore from './stores/historyStore';
+import layoutStore from './stores/layoutStore';
+import menubarStore from './stores/menubarStore';
+import nanoStore from './stores/nanoStore';
+import pathStore from './stores/pathStore';
+import patternStore from './stores/patternStore';
+import pulseStore from './stores/pulseStore';
+import rollStore from './stores/rollStore';
+import sceneStore from './stores/sceneStore';
 
 const stores = {
   layoutStore,
@@ -48,34 +47,6 @@ window.SIREN = stores;
 
 promiseFinally.shim();
 useStrict(false);
-
-// window.onLoad = function() {
-  
-// };
-// document.addEventListener("DOMContentLoaded", function(event) {
-//   console.log("DOM fully loaded and parsed");
-//   var x = document.getElementsByTagName("CANVAS");
-//   var i;
-//   for (i = 0; i < x.length; i++) {
-//     console.log(" -> ", x[i]);
-//     // x[i].style.backgroundColor = "red";
-//   }
-// });
-// document.addEventListener('readystatechange', () => console.log("SSSSSSSSSSSSSSSS:", document.readyState));
-// window.onload = () => { 
-//   console.log("DDDDDDDDDDDDD Window is loaded");
-//   // modules
-//   var x = document.getElementsByClassName('PanelAdjuster');
-//   var i;
-//   for (i = 0; i < x.length; i++) {
-//     console.log(" -> ", x[i]);
-//     x[i].addEventListener('visibilitychange', (state) => {
-//       console.log(state);
-      
-//     });
-//     // x[i].style.backgroundColor = "red";
-//   }
-// };
 
 ReactDOM.render((
   <Provider {...stores}>
