@@ -1,20 +1,23 @@
-import { observable, action, computed } from 'mobx';
+import {
+    observable,
+    action,
+    computed
+} from 'mobx';
 import io from 'socket.io-client';
 import _ from 'lodash';
 
-class DebugStore 
-{
-    sc_log = io('http://localhost:4003/');    
+class DebugStore {
+    sc_log = io('http://localhost:4003/');
 
     @observable msg = '';
-    
+
     constructor() {
         const ctx = this;
         this.sc_log.on('connect', (reason) => {
-            
+
         });
         this.sc_log.on('disconnect', action((reason) => {
-            
+
         }));
         this.sc_log.on("/scdebuglog", action((data) => {
             console.log(data.msg);
@@ -22,17 +25,17 @@ class DebugStore
         }))
     }
 
-    @action updateLog(msg){
+    @action updateLog(msg) {
         var console_len = 5000;
         this.msg = this.msg + msg;
-        if(this.msg.length > console_len){
+        if (this.msg.length > console_len) {
             this.msg = _.drop(this.msg, console_len);
         }
-        
+
     }
-    @computed get debugLogMessage(){
+    @computed get debugLogMessage() {
         return this.msg;
     }
-    
+
 }
 export default new DebugStore();
