@@ -1,5 +1,5 @@
 import axios from 'axios';
-       axios.defaults.timeout = 50000;
-       axios.defaults.baseURL = "http://localhost:3001/";
+axios.defaults.timeout = 50000;
+axios.defaults.baseURL = "http://localhost:3001/";
 
 export default axios;
